python/requirements.txt
kubernetes>=23.0.0
redis>=4.3.0
pyyaml>=6.0
asyncio>=3.4.3
aiohttp>=3.8.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
kopf>=1.35.0  # Optional: Higher-level Kubernetes Operator framework
